tf2/linear_math: fixed quatRotate function
the function was not directly supported by Eigen and wrongly translated
before. Now the code of the Quaternion-class of ROS is used to calculate
the rotation.

TODO: it should be possible to implement this with methods provided by
Eigen itself, but this works for now (might be only implemented in the
rotation matrix and not in the quaternion).