tf2: removed unused files
file that existed in the ROS version but are not needed here.
Mostly implementations for datatypes which are not needed since Eigen is
usedtf2: removed unused files

file that existed in the ROS version but are not needed here.
Mostly implementations for datatypes which are not needed since Eigen is
used