operator/reg_local_icp: first ICP implementation:
- can change data or add a tf

Also a test for ICP is written where the bunny is loaded twice, once
"normal" and once "transformed" with a given tf. On these data the ICP
is performed, which adds a new tf. This tf is loaded with the
buffer_core and compared to the transform used to transform the bunny
originally.

next tasks:
- cleanup, everthing is implemented in one method, this should be
  cleaned up and put into a class.
- support multible inputs at once
- support tf change
- support config values for the ICP