operator/copy: fixed copying of trees
before only a copy of the tree-object was stored at the new path.
This breaks the mapit workspace when the tree has at least one child in the
workspace. In other words, this would only work when every child is only
in the object store and not in the workspace.

Now a deapthFirstSearch is executed starting at the tree and every
entity is copied to there new path. The creation of the new tree is
handeld internaly by the checkout and does not bother the user anymore.