operator/transform: operator to transform data to different frames
when finished this operator should be able to transform all data to
different frames. Currently only pointclouds are supported.

To be loseless when transforming pointclouds the pcl::PCLPointCloud2
data type is transformed directly without any conversion to
pcl::PointCloud. For this the transformPointCloud method from pcl_ros
was ported to mapit.