core/serialization: fixed getStreamProvider for writing of commited data
Before only a reading instance of entity data could be requested, this is
wrong. The AbstractEntitydataProvider (which is returned by
getStreamProvider) contains two streams, one for reading and one for
writing. Till now they where always set to the same. In the case that a
read/write instance is requested for an entity data that dos not exists
in the transient storage and only in the persistent storage the
getStreamProvider needs to return an AbstractEntitydataProvider that
reads from the persistent (aka commit) data and writes to the
transistent (aka workspace) data.

Drawback: the AbstractEntitydataProvider should only be used until a
write is operated, if a read after a write is needed a new
AbstractEntitydataProvider should be used. Otherwise, if we get an
AbstractEntitydataProvider with two different streams, we would read the
old data again, and not the newly written data.
  This could be fixed by resetting the read-stream to the write stream
location when a write is finished, but this would have to be implemented
in every EntityType and will probalby forgotten at one EntityType which
would result in incosistent behaviour for different EntityTypes.
I do not know how I could enforce this by the interface :( (yet?).