operator/filter_voxelgrid: fixed operator for huge point clouds
use an octree to split pointclouds into smaller clouds and execute the
voxelgrid filter on them to avoid the "[pcl::VoxelGrid::applyFilter]
Leaf size is too small for the input dataset. Integer indices would
overflow." error of the PCL.

also changed the operator name from voxelgridfilter to filter_voxelgrid
on the way.