operator/filter_voxelgrid: speedup concatination of PCLPointCloud2
when two PCLPointCloud2 are concatenated, the whole memory is copied.
When you have an operation like this where you have a growing
PCLPointCloud2 where a new snipped is added sequently this slows down
enormously, since the growing point cloud is copied over and over.

Therfore this operator now adds 100 (I know static value is not the
best) clouds into one "sub cloud", which is an intermediate cloud that
stores the voxelgrid-result of up to 100 octree-leafs. Then the next
"sub cloud" is started.
Afterwards these "sub clouds" are merged in an binary fashion. This
results in always two clouds of about the same sice been merged, and
reduces the amount of copied data.