tool/gui: add singleton wrapper for tf_buffer
the tf_buffer takes a long time to get created therefore only one should
be used. For this a new class that wraps one tf_buffer that can be used
for QML Entities was created.

It is "mostly" threadsafe ;)
To get it thread save the method get_buffer_core_ro needs to be removed
and occurances where this method where used need to be wrapped directly.
Like e.g. lookupTransform